feat: #fix: 버그 수정 및 코드리뷰 반영- 충돌 해결, size()를 이용한 개수 조회, 쿼리오타수정

+++ src/main/java/com/umc/withme/dto/meet/MeetShortInfoResponse.java
@@ -36,9 +36,9 @@ public class MeetShortInfoResponse {
     private LocalDateTime createdAt;
 
     @Schema(description = "찜 수")
-    private Long meetLikeCount;
+    private Integer meetLikeCount;
 
-    public static MeetShortInfoResponse of(MeetDto meetDto, MemberDto memberDto, Long meetLikeCount) {
+    public static MeetShortInfoResponse of(MeetDto meetDto, MemberDto memberDto, Integer meetLikeCount) {
 
         return new MeetShortInfoResponse(
                 meetDto.getMeetId(),
